show stacking end of cycle date
